did build work

+++ server/controller.js
@@ -1,3 +1,6 @@
+require('dotenv').config();
+const stripe = require('stripe')(process.env.STRIPE_SECRET);
+
 module.exports = {
     getCoffees: (req, res) => {
         const db = req.app.get('db');
@@ -89,10 +92,10 @@ module.exports = {
     }, 
 
     deleteCoffee: (req, res) => {
-        const db = req.app.get('db');
         let {coffee_id} = req.params;
-
-        db.delete_coffee({coffee_id})
+        let {user_id} = req.session.user;
+        const db = req.app.get('db');
+        db.delete_coffee({coffee_id, user_id})
         .then(coffees => {
             res.status(200).send(coffees);
         })
@@ -134,6 +137,41 @@ module.exports = {
         let {user_id} = req.session.user;
         const db = req.app.get('db');
         db.delete_from_cart({cart_id, user_id})
+        .then(cart => {
+            res.status(200).send(cart);
+        })
+        .catch(err => {
+            console.log(err);
+            res.status(500).send(err)
+        })
+    }, 
+
+    handlePayment: (req, res) => {
+        const {amount, token:{id}} = req.body
+        stripe.charges.create(
+            {
+                amount: amount,
+                currency: "usd",
+                source: id,
+                description: "Test charge from Nathan"
+            },
+            (err, charge) => {
+                if(err) {
+                    console.log(err)
+                    return res.status(500).send(err)
+                } else {
+                    console.log(charge)
+                    return res.status(200).send(charge)
+                }
+            }
+        )
+    },
+
+    clearCart: (req,res) => {
+        let {cart_id} = req.params;
+        let {user_id} = req.session.user;
+        const db = req.app.get('db');
+        db.clear_cart({cart_id, user_id})
         .then(cart => {
             res.status(200).send(cart);
         })
